+++ frontend/src/router/index.js
@@ -1,6 +1,5 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import Home from '../views/Home'
-import Test from '../views/Test'
 
 const routes = [
   {
@@ -8,11 +7,6 @@ const routes = [
     name: 'Home',
     component: Home,
   },
-  {
-    path: '/test',
-    name: 'Test',
-    component: Test,
-  }
 ]
 
 const router = createRouter({
